refactor(glyph): extract interpolation quasi rendering into a submodule
The source-length ratchet fails when an already-over-limit file grows
versus the PR base. Splitting the interpolation-quasi logic and its tests
out keeps template.rs and formatter.rs at or below their base line counts
while preserving behaviour.

- Move render_interpolation_expr_lines and template_literal_quasi_line_starts
  into template/formatter/interpolation.rs (child module of formatter, so it
  keeps access to the private indent/newline helpers).
- Move the three interpolation idempotence tests from template.rs's inline
  test module into tests/template_literal_interpolation.rs, driven through the
  public format_template API.
